Refactor: Centralisation des types des composants de fonctionnalités
