+++ src/components/features/index.ts
@@ -3,8 +3,3 @@ export { default as FeatureCard } from './FeatureCard';
 export { default as FeatureIcon } from './FeatureIcon';
 export { default as DetailedFeature } from './DetailedFeature';
 export { default as DetailedFeaturesSection } from './DetailedFeaturesSection';
-
-export type { FeatureCardProps } from './FeatureCard';
-export type { FeatureIconProps } from './FeatureIcon';
-export type { DetailedFeatureProps } from './DetailedFeature';
-export type { DetailedFeaturesSectionProps } from './DetailedFeaturesSection';
